firefox版に軽量の時間割を追加
