+++ src/contents/components/firefoxWidgetStyle.module.css
@@ -1,0 +1,120 @@
+.utlContainer {
+  position: fixed;
+  top: 0;
+  right: 10px;
+  width: calc(100% - 320px);
+  height: 100%;
+  background-color: #0000;
+  z-index: 9999;
+  color: #000;
+}
+@media (max-width: 700px) {
+  .utlContainer {
+    display: none;
+  }
+}
+
+.widgetContainer {
+  width: calc(100% - 20px);
+  display: block;
+  background-color: #fff7;
+  backdrop-filter: blur(6px);
+  padding: 8px;
+  border-radius: 4px;
+  margin: 10px auto 0 auto;
+}
+
+.timetableBody {
+  display: grid;
+  width: calc(100% - 4px);
+  background-color: #eef7f799;
+  column-gap: 2px;
+  row-gap: 5px;
+  padding: 2px;
+  border-radius: 5px;
+}
+
+.timetablePeriod {
+  text-align: center;
+  background-color: #eda49c;
+  font-size: 90%;
+}
+
+.weekdayHeader {
+  text-align: center;
+  background-color: #f0d6a0;
+}
+
+.timetableRowHeader {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  gap: 0;
+  background: #eda49c;
+}
+
+.timetablePeriodTime {
+  font-size: 10px;
+  display: block;
+  color: gray;
+  line-height: 120%;
+}
+
+.timetableCell {
+  display: flex;
+  flex-direction: column;
+  gap: 3px;
+  border: 1px solid #bbb;
+  padding: 2px;
+  border: 1px solid #999;
+}
+
+.timetableClassroom {
+  line-height: 130%;
+  letter-spacing: 0;
+  display: inline-block;
+}
+
+.timetableNowClass {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(91, 237, 146);
+  opacity: 0.3;
+}
+
+.timetableButton {
+  text-align: center;
+  box-sizing: border-box;
+  padding: 3px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 36px;
+  overflow: hidden;
+  gap: 3px;
+  position: relative;
+  box-sizing: border-box;
+  color: #000;
+  background-color: #fff;
+  border-radius: 5px;
+  margin: 2px;
+  cursor: pointer;
+  box-shadow: 0 0 5px 0 rgba(0, 0, 0, 0.2);
+}
+.timetableButton:hover {
+  opacity: 0.7;
+  border: 1px solid #000;
+  padding: 2px;
+}
+
+.timetableTopDate {
+  font-size: 16px;
+  text-align: center;
+  margin: 2px 0 7px 0;
+  font-weight: normal;
+}
