+++ src/options/container/firefox.tsx
@@ -1,0 +1,10 @@
+const OptionsIndexFirefox = () => {
+  return (
+    <div>
+      <h1>Options</h1>
+      <p>Options for Firefox</p>
+    </div>
+  );
+};
+
+export default OptionsIndexFirefox;
